service/mqtt_to_rdf/requirements.txt
cyclone
rdflib-jsonld==0.5.0
rdflib==4.2.2
service_identity==18.1.0
twisted-mqtt==0.3.9
rx==3.1.1
docopt
prometheus_client==0.8.0

cycloneerr
export_to_influxdb==0.4.0
mqtt_client==0.9.0
patchablegraph==0.11.0
rdfdb==0.21.0
standardservice==0.6.0